+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+### Added
+- Task priority support with setPriority() method
+
 ## [1.3.0] - 2024-XX-XX
 ### Added
 - Task priority support with setPriority() method